+++ changelog.md
@@ -7,6 +7,7 @@
 * Modified AppInsights configuration, added to the Vue3 migration, and updated to latest version
 * Added settings configuration support to the Vue3 port
 * Update CSIRO renderer dependencies:  sdc-populate@4.7.0, sdc-template-extract@1.0.15, @aehrc/smart-forms-renderer@1.3.1
+* Update the FML rendering diagrams to resolve the type definitions! (and show red backgrounds if the property doesn't resolve)
 
 # 24 April 2026
 * Update NLM LHC-Forms to 41.2.0